+++ core/package.json
@@ -23,9 +23,6 @@
     "prepare": "npm run build",
     "test": "jest"
   },
-  "dependencies": {
-    "keyborg": "^1.0.0"
-  },
   "//": "core-js needs to be present for storybook: https://github.com/storybookjs/storybook/blob/master/MIGRATION.md#core-js-dependency-errors",
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.12.1",
