+++ core/src/State/KeyboardNavigation.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { createKeyborg, disposeKeyborg, Keyborg } from 'keyborg';
+import { createKeyborg, disposeKeyborg, Keyborg } from '../Keyborg';
 
 import { Subscribable } from './Subscribable';
 import * as Types from '../Types';
