--- keyborg/.gitignore
@@ -1,2 +1,0 @@
-/dist
-/.rts2*
