--- keyborg/.npmignore
@@ -1,2 +1,0 @@
-/tsconfig.json
-/tslint.json
