--- keyborg/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "keyborg",
-  "version": "1.0.0",
-  "description": "Keyboard Navigation Detection for Web",
-  "author": "Marat Abdullin <marata@microsoft.com>",
-  "license": "MIT",
-  "sideEffects": false,
-  "main": "./dist/index.js",
-  "module": "./dist/keyborg.esm.js",
-  "typings": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/microsoft/tabster"
-  },
-  "scripts": {
-    "build": "npm run tslint && npm run clean && tsdx build",
-    "clean": "rimraf dist",
-    "tslint": "tslint --project tsconfig.json --fix || true",
-    "tslint:check": "tslint --project tsconfig.json",
-    "prepare": "npm run build"
-  },
-  "devDependencies": {
-    "lint": "^1.0.0-alpha.0",
-    "rimraf": "^3.0.2",
-    "tsdx": "^0.14.1",
-    "tslint": "^6.1.3",
-    "tsutils": "^3.21.0",
-    "typescript": "^4.2.3"
-  }
-}
