--- keyborg/tsconfig.json
@@ -1,32 +1,0 @@
-{
-    "compilerOptions": {
-        "declaration": true,
-        "noResolve": false,
-        "module": "es2015",
-        "target": "es5",
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noImplicitThis": true,
-        "noUnusedLocals": true,
-        "skipLibCheck": true,
-        "strictNullChecks": true,
-        "strictPropertyInitialization": true,
-        "forceConsistentCasingInFileNames": true,
-        "outDir": "./dist/",
-        "lib": [
-            "es5",
-            "es2015.iterable",
-            "dom",
-            "esnext"
-        ]
-    },
-
-    "include": [
-        "./src/**/*"
-    ],
-
-    "exclude": [
-        "dist",
-        "node_modules"
-    ]
-}
