--- keyborg/tslint.json
@@ -1,3 +1,0 @@
-{
-    "extends": ["lint"]
-}
